src/test/javascript/spec/app/entities/beer/beer-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';
import { ToastPlugin } from 'bootstrap-vue';

import * as config from '@/shared/config/config';
import BeerUpdateComponent from '@/entities/beer/beer-update.vue';
import BeerClass from '@/entities/beer/beer-update.component';
import BeerService from '@/entities/beer/beer.service';

import AlertService from '@/shared/alert/alert.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.use(ToastPlugin);
localVue.component('font-awesome-icon', {});
localVue.component('b-input-group', {});
localVue.component('b-input-group-prepend', {});
localVue.component('b-form-datepicker', {});
localVue.component('b-form-input', {});

describe('Component Tests', () => {
  describe('Beer Management Update Component', () => {
    let wrapper: Wrapper<BeerClass>;
    let comp: BeerClass;
    let beerServiceStub: SinonStubbedInstance<BeerService>;

    beforeEach(() => {
      beerServiceStub = sinon.createStubInstance<BeerService>(BeerService);

      wrapper = shallowMount<BeerClass>(BeerUpdateComponent, {
        store,
        localVue,
        router,
        provide: {
          beerService: () => beerServiceStub,
          alertService: () => new AlertService(),
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.beer = entity;
        beerServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(beerServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.beer = entity;
        beerServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(beerServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        const foundBeer = { id: 123 };
        beerServiceStub.find.resolves(foundBeer);
        beerServiceStub.retrieve.resolves([foundBeer]);

        // WHEN
        comp.beforeRouteEnter({ params: { beerId: 123 } }, null, cb => cb(comp));
        await comp.$nextTick();

        // THEN
        expect(comp.beer).toBe(foundBeer);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        comp.previousState();
        await comp.$nextTick();

        expect(comp.$router.currentRoute.fullPath).toContain('/');
      });
    });
  });
});
